Moved code to common
