+++ build.gradle
@@ -220,7 +220,7 @@ dependencies {
   compile group: 'uk.gov.hmcts.reform', name: 'send-letter-client', version: '2.6.0'
   compile group: 'uk.gov.hmcts.reform', name: 'service-auth-provider-client', version: '3.1.2'
 
-  compile group: 'uk.gov.hmcts.reform', name: 'sscs-common', version: '3.4.52'
+  compile group: 'uk.gov.hmcts.reform', name: 'sscs-common', version: '3.4.59-RC-SSCS-7860'
   compile group: 'uk.gov.hmcts.reform', name: 'sscs-pdf-email-common', version: '1.4.9'
 
   //Remove when our dependencies pull in this version or later
