--- src/main/resources/schema/sscs-robotics.json
@@ -1,187 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema",
-  "title": "Case",
-  "type": "object",
-  "properties": {
-    "caseCode": {
-      "type": "string",
-      "pattern": "^\\d{3}[A-Z]{2}$"
-    },
-    "caseId" : {"type" : "integer"},
-    "appellantNino":  {"type": "string"},
-    "appellantPostCode": {"type": "string"},
-    "evidencePresent": {"$ref": "#/definitions/yesOrNo"},
-    "appealDate": {"$ref": "#/definitions/dateType"},
-    "mrnDate": {"$ref": "#/definitions/dateType"},
-    "mrnReasonForBeingLate": {"type": "string"},
-    "hearingType": {
-      "type": "string",
-      "enum": ["Paper", "Oral"]
-    },
-    "hearingRequestParty": {"type" : "string"},
-    "pipNumber" : {"type" : "string"},
-    "receivedVia" : {"type" : "string"},
-    "appellant": {"$ref": "#/definitions/person"},
-    "appointee": {"$ref": "#/definitions/person"},
-    "representative": {"$ref": "#/definitions/person"},
-    "jointParty": {"$ref": "#/definitions/person"},
-    "rpcEmail": {"type": "string"},
-    "isReadyToList": {"$ref": "#/definitions/yesOrNo"},
-    "isDigital": {"$ref": "#/definitions/yesOrNo"},
-    "dwpResponseDate": {"$ref": "#/definitions/dateType"},
-    "dwpIssuingOffice": {"type": "string"},
-    "dwpPresentingOffice": {"type": "string"},
-    "dwpIsOfficerAttending": {"$ref": "#/definitions/yesOrNo"},
-    "dwpUcb": {"$ref": "#/definitions/yesOrNo"},
-    "hearingArrangements": {
-      "type": "object",
-      "properties" : {
-        "languageInterpreter": {"type" : "string"},
-        "signLanguageInterpreter": {"type" : "string"},
-        "hearingLoop": {"$ref": "#/definitions/yesOrNo"},
-        "accessibleHearingRoom": {"$ref": "#/definitions/yesOrNo"},
-        "other": {"type" : "string"},
-        "datesCantAttend": {
-          "type" : "array",
-          "items": {
-            "$ref": "#/definitions/dateType"
-          }
-        }
-      },
-      "required": [
-        "hearingLoop",
-        "accessibleHearingRoom"
-      ],
-      "additionalProperties": false
-    },
-    "ucDecisionDisputedByOthers": {"$ref": "#/definitions/yesOrNo"},
-    "linkedAppealRef": {"type": "string"},
-    "elementsDisputed": {
-      "type": "object",
-      "properties": {
-        "general": {
-          "type": "array",
-          "items": {
-            "type": "string"
-          }
-        },
-        "sanctions": {
-          "type": "array",
-          "items": {
-            "type": "string"
-          }
-        },
-        "overpayment": {
-          "type": "array",
-          "items": {
-            "type": "string"
-          }
-        },
-        "housing": {
-          "type": "array",
-          "items": {
-            "type": "string"
-          }
-        },
-        "childCare": {
-          "type": "array",
-          "items": {
-            "type": "string"
-          }
-        },
-        "care": {
-          "type": "array",
-          "items": {
-            "type": "string"
-          }
-        },
-        "childElement": {
-          "type": "array",
-          "items": {
-            "type": "string"
-          }
-        },
-        "childDisabled": {
-          "type": "array",
-          "items": {
-            "type": "string"
-          }
-        }
-      }
-    }
-  },
-  "dependencies": {
-    "hearingType": {
-      "oneOf": [
-        {
-          "properties": {
-            "hearingType": {
-              "enum": ["Paper"]
-            }
-          },
-          "not": { "required": ["hearingRequestParty"] }
-        },
-        {
-          "properties": {
-            "hearingType": {
-              "enum": ["Oral"]
-            }
-          },
-          "required": ["hearingRequestParty"]
-        }
-      ]
-    }
-  },
-
-  "additionalProperties": false,
-  "required": [
-    "caseCode",
-    "appellantNino",
-    "appealDate",
-    "hearingType",
-    "appellant",
-    "evidencePresent"
-  ],
-
-  "definitions": {
-    "person": {
-      "type": "object",
-      "properties": {
-        "title" : { "type" : "string"},
-        "firstName": { "type": "string" },
-        "lastName": { "type": "string"},
-        "organisation": { "type": "string"},
-        "addressLine1": { "type": "string"},
-        "addressLine2": { "type": "string"},
-        "townOrCity": { "type": "string"},
-        "county": { "type": "string" },
-        "postCode": {
-          "type": "string",
-          "pattern": "^((([A-Za-z][0-9]{1,2})|(([A-Za-z][A-Ha-hJ-Yj-y][0-9]{1,2})|(([A-Za-z][0-9][A-Za-z])|([A-Za-z][A-Ha-hJ-Yj-y][0-9]?[A-Za-z])|([Gg][Ii][Rr]))))\\s?([0-9][A-Za-z]{2})|(0[Aa]{2}))$"},
-        "phoneNumber": { "type": "string" },
-        "email": { "type": "string" },
-        "dob": {"$ref": "#/definitions/dateType"},
-        "nino": { "type": "string" },
-        "sameAddressAsAppellant": {"$ref": "#/definitions/yesOrNo"}
-      },
-      "additionalProperties": false,
-      "required": [
-        "firstName",
-        "lastName",
-        "addressLine1",
-        "townOrCity",
-        "county",
-        "postCode"
-      ]
-    },
-    "yesOrNo" : {
-      "type": "string",
-      "enum": ["Yes", "No"]
-    },
-    "dateType": {
-      "type": "string",
-      "format": "date"
-    }
-  }
-}
-
